feat: refactored clients table
