+++ next/src/types/client-types.ts
@@ -1,0 +1,16 @@
+export type ClientsTypes = {
+  id: string;
+  avatar: string;
+  birthday: Date;
+  email: string;
+  firstName: string;
+  lastName: string;
+  sex: string;
+  supportTier: 'standard' | 'gold' | 'platinum';
+  hourlyRate: number;
+};
+
+export type ClientTableHeaderRow = {
+  title: string;
+  classes?: string;
+};
